chores: refactor some pkg dir names
