+++ pkg/controller/metrics/yarn_collector.go
@@ -14,12 +14,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package noderesource
+package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/koordinator-sh/goyarn/pkg/yarn/cache"
+)
+
+const (
+	yarnNodeCPUResource             = "yarn_node_cpu_resource"
+	yarnNodeMemoryResource          = "yarn_node_memory_resource"
+	yarnNodeCPUAllocatedResource    = "yarn_node_cpu_allocated_resource"
+	yarnNodeMemoryAllocatedResource = "yarn_node_memory_allocated_resource"
 )
 
 var (
