+++ pkg/copilot/nm/nm_pod_discover.go
